requirements.txt
# API
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Core ML stack
torch>=2.2
numpy>=1.26
pandas>=2.2
scikit-learn>=1.5
matplotlib>=3.9

# DB
SQLAlchemy>=2.0
psycopg2-binary>=2.9
alembic>=1.13

# Utils
pydantic>=2.7
python-dotenv>=1.0
PyYAML>=6.0
loguru>=0.7
croniter>=3.0